[srs] Remove unneeded factory_key type.
